requirements.txt
fastapi
uvicorn
requests
pillow
pandas
python-multipart
gspread
google-auth
